changeMessage() is now a context-free grammar
This allows other string patterns to be added easily, for example replacing "the" with a choice of other modifiers.

Also, the random percentage case is calculated each time it is needed, instead of once when the page loads.

New tags should be added to the grammar object.  The key should be a word, and the value should be a list of strings, some of which may contain other keys in the grammar.  Be careful; there is no check for infinite recursion.  Also, if the same tag appears twice at any stage of processing, both tags will be replaced with the same choice.